+++ albumart.php
@@ -65,7 +65,7 @@ include ("includes/globals.php");
 print '<tr><td colspan="2"><h2>'.get_int_text("albumart_title").'</h2></td><td class="outer" align="right"><button id="doobag">'.get_int_text("albumart_findsmall").'</button></td><td class="outer" align="right"><button id="harold">'.get_int_text("albumart_getmissing").'</button></td></tr>';
 print '<tr><td class="outer" id="totaltext"></td><td colspan=""><div class="invisible" id="progress"></div></td><td class="outer styledinputs" align="right"><input type="checkbox" class="topcheck" id="dinkytoys"><label for="dinkytoys" onclick="toggleLocal()">Ignore Local Images</label></td>
 <td class="outer styledinputs" align="right"><input type="checkbox" class="topcheck" id="poobag"><label for="poobag" onclick="toggleScrolling()">Follow Progress</label></td></tr>';
-print '<tr><td class="outer" id="infotext"></td><td colspan="2" align="center"><div class="inner" id="status">Loading...</div></td><td class="outer" align="right"><button id="finklestein">'.get_int_text("albumart_onlyempty").'</button></td></tr>';
+print '<tr><td class="outer" id="infotext"></td><td colspan="2" align="center"><div class="inner" id="status">'.get_int_text('label_loading').'</div></td><td class="outer" align="right"><button id="finklestein">'.get_int_text("albumart_onlyempty").'</button></td></tr>';
 ?>
 </table>
 </div>
@@ -123,7 +123,7 @@ function do_covers_db_style() {
     foreach ($alist as $artist) {
         print '<div class="cheesegrater" name="artistname'.$artist['Artistindex'].'">';
             print '<div class="albumsection">';
-                print '<div class="tleft"><h2>'.$artist['Artistname'].'</h2></div><div class="tright rightpad"><button onclick="getNewAlbumArt(\'#album'.$count.'\')">'.get_int_text("albumart_getthese").'</button></div>';
+                print '<div class="tleft"><h2>'.$artist['Artistname'].'</h2></div><div class="tright rightpad"><button class="invisible" onclick="getNewAlbumArt(\'#album'.$count.'\')">'.get_int_text("albumart_getthese").'</button></div>';
             print "</div>\n";
                 print '<div id="album'.$count.'" class="containerbox fullwidth bigholder wrap">';
                 $blist = get_list_of_albums($artist['Artistindex']);
@@ -165,7 +165,7 @@ function do_radio_stations() {
     if (count($playlists) > 0) {
         print '<div class="cheesegrater" name="radio">';
             print '<div class="albumsection">';
-                print '<div class="tleft"><h2>Radio Stations</h2></div><div class="tright rightpad"><button onclick="getNewAlbumArt(\'#album'.$count.'\')">'.get_int_text("albumart_getthese").'</button></div>';
+                print '<div class="tleft"><h2>Radio Stations</h2></div><div class="tright rightpad"><button class="invisible" onclick="getNewAlbumArt(\'#album'.$count.'\')">'.get_int_text("albumart_getthese").'</button></div>';
                 print "</div>\n";
                 print '<div id="album'.$count.'" class="containerbox fullwidth bigholder wrap">';
                 foreach ($playlists as $file) {