+++ includes/prefspanel.php
@@ -29,21 +29,23 @@ foreach($themes as $theme) {
 print '</select></div></div>';
 
 // Custom Background
-print '<div id="custombackground" class="pref containerbox dropdown-container">
-<div class="divlabel">'.get_int_text('config_background').'
-<div id="cusbgname" class="tiny" style="font-weight:normal"></div>
-</div>
-<div class="selectholder-noselect">
+print '<div id="custombackground" class="pref containerbox">
+<div id="backimagelabel" class="divlabel">'.get_int_text('config_background').'
 <form id="backimageform" action="backimage.php" method="post" enctype="multipart/form-data">
 <input type="hidden" name="currbackground" value="" />
 <input type="hidden" name="browser_id" value="" />
-<input type="file" name="imagefile" class="infowiki">
-<div class="pref styledinputs">
+<div class="filebutton textcentre">
+<input type="file" name="imagefile[]" id="imagefile" class="inputfile" multiple="multiple">
+<label for="imagefile">'.get_int_text('label_choosefiles').'</label>
+</div>
+<input type="button" class="invisible" onclick="prefs.changeBackgroundImage()" value="'.get_int_text('albumart_uploadbutton').'">
+<div class="styledinputs">
 <input type="checkbox" id="thisbrowseronly" name="thisbrowseronly" /><label for="thisbrowseronly">For this browser only</label>
 </div>
-<input type="button" onclick="prefs.changeBackgroundImage()" value="'.get_int_text('albumart_uploadbutton').'">
-<i class="icon-cancel-circled clickicon collectionicon" onclick="prefs.clearBgImage()"></i>
 </form>
+</div>
+<div class="selectholder-noselect">
+<div id="cusbgname" class="tiny styledinputs"></div>
 </div>
 </div>';
 
@@ -248,13 +250,6 @@ print '<div class="pref styledinputs">
 <input class="saveotron prefinput" id="music_directory_albumart" type="text" size="40" />
 </div>';
 print '<div class="tiny">If you want to use Google Images to get Album Art you need to sign up for an API Key. Please <a href="https://fatg3erman.github.io/RompR/Album-Art-Manager" target="_blank">'.get_int_text('config_read_the_docs').'</a></div>';
-// print '<div class="pref"><b>'.get_int_text('config_google_credentials').'</b></div>
-// <div class="pref"><a href="https://fatg3erman.github.io/RompR/Album-Art-Manager" target="_blank">'.get_int_text('config_read_the_docs').'</a></div>';
-// print '<div class="pref"><b>Google API Key</b>
-// <input class="saveotron prefinput" id="google_api_key" type="text" size="120" /></div>
-// <div class="pref"><b>Google Search Engine ID</b>
-// <input class="saveotron prefinput" id="google_search_engine_id" type="text" size="120" /></div>';
-
 
 print '<div class="textcentre configtitle">
 <i class="icon-podcast-circled medicon"></i><b>'.get_int_text('label_podcasts').'</b></div>';