+++ streamplugins/01_somafm.php
@@ -22,7 +22,7 @@ class somafmplugin {
             'expand' => true
         ));
         print '<div id="somafmlist" class="dropmenu notfilled">';
-        print '<div class="textcentre">Loading...</div></div>';
+        print '<div class="configtitle textcentre"><b>'.get_int_text('label_loading').'</b></div></div>';
         print '</div>';
     }