+++ streamplugins/85_iceScraper.php
@@ -131,7 +131,7 @@ if (array_key_exists('populate', $_REQUEST)) {
 		'class' => 'radio',
 		'expand' => true
     ));
-	print '<div id="icecastlist" class="dropmenu notfilled"><div class="textcentre">Loading...</div></div>';
+	print '<div id="icecastlist" class="dropmenu notfilled"><div class="configtitle textcentre"><b>'.get_int_text('label_loading').'</b></div></div>';
 	print '</div>';
 }