+++ ui/prefs.js
@@ -69,7 +69,8 @@ var prefs = function() {
         "browser_id",
         "playlistswipe",
         "podcastcontrolsvisible",
-        "use_albumart_in_playlist"
+        "use_albumart_in_playlist",
+        "bgimgparms"
     ];
     
     const cookiePrefs = [
@@ -81,6 +82,21 @@ var prefs = function() {
 	const jsonNode = document.querySelector("script[name='prefs']");
   	const jsonText = jsonNode.textContent;
   	const tags = JSON.parse(jsonText);
+    
+    var backgroundImages;
+    var backgroundTimer;
+    
+    var timeouts = {
+        '10 Seconds': 10000,
+        '30 Seconds': 30000,
+        'Minute': 60000,
+        '5 Minutes': 300000,
+        '10 Minutes': 600000,
+        '20 Minutes': 1200000,
+        '30 Minutes': 1800000,
+        'Hour': 3600000,
+        'Day': 86400000
+    }
 
     function doTheSave() {
         var felakuti = new Object;
@@ -133,38 +149,148 @@ var prefs = function() {
         prefs.save({scrobblepercent: e.max});
     }
 
+    function loadBackgroundImages(theme) {
+        $('#cusbgname').empty();
+        $.getJSON('backimage.php?getbackground='+theme+'&browser_id='+prefs.browser_id, function(data) {
+            debug.log("PREFS","Custom Background Image",data);
+            if (data.images) {
+                if (typeof(prefs.bgimgparms[theme]) == 'undefined') {
+                    debug.trace("PREFS","Init bgimgparms for",prefs.theme);
+                    prefs.bgimgparms[theme] = {
+                        landscape: 0,
+                        portrait: 0,
+                        timeout: 60000,
+                        lastchange: Date.now(),
+                        random: false
+                    }
+                    prefs.save({bgimgparms: prefs.bgimgparms});
+                }
+                setCustombackground(data.images);
+                $('input[name="thisbrowseronly"]').prop('checked', data.thisbrowseronly);
+            }
+        });
+    }
+    
+    function changeRandomMode() {
+        var theme = prefs.theme;
+        if (prefs.usertheme) {
+            theme = prefs.usertheme;
+        }
+        if (typeof(prefs.bgimgparms[theme].random) == 'undefined') {
+            prefs.bgimgparms[theme].random = false;
+        }
+        prefs.bgimgparms[theme].random = !prefs.bgimgparms[theme].random;
+        prefs.save({bgimgparms: prefs.bgimgparms});
+    }
+    
+    function removeAllBackgroundImages() {
+        clearCustomBackground();
+        var theme = prefs.theme;
+        if (prefs.usertheme) {
+            theme = prefs.usertheme;
+        }
+        $.getJSON('backimage.php?clearallbackgrounds='+theme+'&browser_id='+prefs.browser_id, function(data) {
+            loadBackgroundImages(theme);
+        });
+    }
+
     function setCustombackground(images) {
+        clearTimeout(backgroundTimer);
         debug.log("UI","Setting Custom Background To",images);
+        backgroundImages = images;
+        if (images.landscape.length > 1 || images.portrait.length > 1) {
+            var jesus = $('<div>', {class: 'containerbox dropdown-container'}).appendTo('#cusbgname');
+            jesus.append('<div class="divlabel">'+language.gettext('label_changevery')+'</div>');
+            var sh = $('<div>', {class: 'selectholder'}).appendTo(jesus);
+            var s = $('<select>').appendTo(sh);
+            $.each(timeouts, function(i, v){
+                s.append('<option value="'+v+'">'+i+'</option>');
+            });
+            s.val(prefs.bgimgparms[prefs.theme].timeout.toString());
+            s.on('change', function() {
+                prefs.bgimgparms[prefs.theme].timeout = parseInt(s.val());
+                prefs.save({bgimgparms: prefs.bgimgparms});
+                updateCustomBackground();
+            });
+            var gibbon = $('<div>', {class: 'clearfix'}).appendTo('#cusbgname');
+            var ran = $('<input>', {type: 'checkbox', id: 'bgimagerandom'}).appendTo(gibbon);
+            var lab = $('<label>', {for: 'bgimagerandom'}).appendTo(gibbon);
+            lab.html('Random Order');
+            ran.prop('checked', prefs.bgimgparms[prefs.theme].random);
+            ran.bind('click', changeRandomMode);
+            var rb = $('<button>', {class: 'tright'}).appendTo(gibbon);
+            rb.html('Remove All');
+            rb.bind('click', removeAllBackgroundImages);
+        }
+        $.each(images, function(x, p) {
+            $.each(p, function(i, v) {
+                var n = $('<div>').appendTo('#cusbgname');
+                var c = $('<i>', {class: 'icon-cancel-circled clickicon collectionicon'}).appendTo(n);
+                var l = $('<span>', {class: 'bgimgname'}).appendTo(n);
+                var x = $('<input>', {class: 'bgimagefile', type: 'hidden', value: v}).appendTo(n);
+                l.html(v.replace(/.*(\\|\/)/, ''));
+                c.bind('click', prefs.clearBgImage);
+            });
+        });
+        updateCustomBackground();
+    }
+
+    function clearCustomBackground() {
         $('style[id="phoneback"]').remove();
         $('style[id="background"]').remove();
         $('style[id="phonebackl"]').remove();
         $('style[id="backgroundl"]').remove();
         $('style[id="phonebackp"]').remove();
         $('style[id="backgroundp"]').remove();
-        var cusname = new Array();
-        if (images != '') {
-            for (var i in images) {
-                if (images[i].match(/_landscape/)) {
-                    $('style[id="phonebackl"]').remove();
-                    $('style[id="backgroundl"]').remove();
-                    $('<style id="backgroundl">@media screen and (orientation: landscape) { html { background-image: url("'+images[i]+'?version='+rompr_version+'"); background-size: cover; background-repeat: no-repeat } }</style>').appendTo('head');
-                    $('<style id="phonebackl">@media screen and (orientation: landscape) { body.phone .dropmenu { background-image: url("'+images[i]+"?version="+rompr_version+'") } }</style>').appendTo('head');
-                } else if (images[i].match(/_portrait/)) {
-                    $('style[id="phonebackp"]').remove();
-                    $('style[id="backgroundp"]').remove();
-                    $('<style id="backgroundp">@media screen and (orientation: portrait) { html { background-image: url("'+images[i]+'?version='+rompr_version+'"); background-size: cover; background-repeat: no-repeat } }</style>').appendTo('head');
-                    $('<style id="phonebackp">@media screen and (orientation: portrait) { body.phone .dropmenu { background-image: url("'+images[i]+"?version="+rompr_version+'") } }</style>').appendTo('head');
-                } else {
-                    $('style[id="phoneback"]').remove();
-                    $('style[id="background"]').remove();
-                    $('<style id="background">html { background-image: url("'+images[i]+'?version='+rompr_version+'"); background-size: cover; background-repeat: no-repeat }</style>').appendTo('head');
-                    $('<style id="phoneback">body.phone .dropmenu { background-image: url("'+images[i]+"?version="+rompr_version+'") }</style>').appendTo('head');
-                }
-                cusname.push(images[i].split(/[\\/]/).pop());
+        clearTimeout(backgroundTimer);
+    }
+
+    function updateCustomBackground() {
+        clearTimeout(backgroundTimer);
+        clearCustomBackground();
+        var theme = prefs.theme;
+        if (prefs.usertheme) {
+            theme = prefs.usertheme;
+        }
+        var bgp = prefs.bgimgparms[theme];
+        var timeout = bgp.timeout;
+        if (bgp.timeout + bgp.lastchange <= Date.now()) {
+            bgp.lastchange = Date.now();
+            if (bgp.random) {
+                bgp.landscape = Math.floor(Math.random() * backgroundImages.landscape.length);
+                bgp.portrait = Math.floor(Math.random() * backgroundImages.portrait.length);
+            } else {
+                bgp.landscape++;
+                bgp.portrait++;
             }
-            $('#cusbgname').html(cusname.join(', '));
+            prefs.save({bgimgparms: prefs.bgimgparms});
         } else {
-            $('#cusbgname').html('');
+            timeout = bgp.timeout + bgp.lastchange - Date.now();
+        }
+        if (bgp.landscape >= backgroundImages.landscape.length) { bgp.landscape = 0 }
+        if (bgp.portrait >= backgroundImages.portrait.length) { bgp.portrait = 0 }
+        if (backgroundImages.portrait.length == 0) {
+            $('<style id="background">html { background-image: url("'+backgroundImages.landscape[bgp.landscape]+'") }</style>').appendTo('head');
+            $('<style id="phoneback">body.phone .dropmenu { background-image: url("'+backgroundImages.landscape[bgp.landscape]+'") }</style>').appendTo('head');
+            $('span.bgimgname').removeClass('selected');
+            $('input.bgimagefile[value="'+backgroundImages.landscape[bgp.landscape]+'"]').prev().addClass('selected');
+        } else if (backgroundImages.landscape.length == 0) {
+            $('<style id="background">html { background-image: url("'+backgroundImages.portrait[bgp.portrait]+'") }</style>').appendTo('head');
+            $('<style id="phoneback">body.phone .dropmenu { background-image: url("'+backgroundImages.portrait[bgp.portrait]+'") }</style>').appendTo('head');
+            $('span.bgimgname').removeClass('selected');
+            $('input.bgimagefile[value="'+backgroundImages.portrait[bgp.portrait]+'"]').prev().addClass('selected');
+        } else {
+            $('<style id="backgroundl">@media screen and (orientation: landscape) { html { background-image: url("'+backgroundImages.landscape[bgp.landscape]+'") } }</style>').appendTo('head');
+            $('<style id="phonebackl">@media screen and (orientation: landscape) { body.phone .dropmenu { background-image: url("'+backgroundImages.landscape[bgp.landscape]+'") } }</style>').appendTo('head');
+            $('<style id="backgroundp">@media screen and (orientation: portrait) { html { background-image: url("'+backgroundImages.portrait[bgp.portrait]+'") } }</style>').appendTo('head');
+            $('<style id="phonebackp">@media screen and (orientation: portrait) { body.phone .dropmenu { background-image: url("'+backgroundImages.portrait[bgp.portrait]+'") } }</style>').appendTo('head');
+            $('span.bgimgname').removeClass('selected');
+            $('input.bgimagefile[value="'+backgroundImages.landscape[bgp.landscape]+'"]').prev().addClass('selected');
+            $('input.bgimagefile[value="'+backgroundImages.portrait[bgp.portrait]+'"]').prev().addClass('selected');
+        }
+        if (backgroundImages.portrait.length > 1 || backgroundImages.landscape.length > 1) {
+            debug.trace("PREFS","Setting Slideshow Timeout For",timeout/1000,"seconds");
+            backgroundTimer = setTimeout(updateCustomBackground, timeout);
         }
     }
 
@@ -463,7 +589,7 @@ var prefs = function() {
 
         setTheme: function(theme) {
             if (!theme) theme = prefs.theme;
-            setCustombackground('');
+            clearCustomBackground();
             // Use a different version every time to ensure the browser doesn't cache.
             // Browsers are funny about CSS.
             var t = Date.now();
@@ -473,13 +599,7 @@ var prefs = function() {
             $("#fontfamily").attr("href", "fonts/"+prefs.fontfamily+"?version="+t);
             $("#icontheme-theme").attr("href", "iconsets/"+prefs.icontheme+"/theme.css"+"?version="+t);
             $("#icontheme-adjustments").attr("href", "iconsets/"+prefs.icontheme+"/adjustments.css"+"?version="+t);
-            $.getJSON('backimage.php?getbackground='+theme+'&browser_id='+prefs.browser_id, function(data) {
-                if (data.images) {
-                    debug.log("PREFS","Custom Background Image",data.images);
-                    setCustombackground(data.images);
-                    $('input[name="thisbrowseronly"]').prop('checked', data.thisbrowseronly);
-                }
-            });
+            loadBackgroundImages(theme);
             prefs.rgbs = null;
             if (typeof(layoutProcessor) != 'undefined') {
                 setTimeout(prefs.postUIChange, 2000);
@@ -517,10 +637,19 @@ var prefs = function() {
             xhr.send(new FormData(formElement));
         },
 
-        clearBgImage: function() {
-            setCustombackground('');
-            $.getJSON('backimage.php?clearbackground='+prefs.theme+'&browser_id='+prefs.browser_id, function(data) {
-                $('[name=imagefile').val('');
+        clearBgImage: function(event) {
+            var clicked = $(event.target);
+            var image = clicked.next().next().val();
+            clicked.parent().remove();
+            clearCustomBackground();
+            $.getJSON('backimage.php?clearbackground='+image, function(data) {
+                $('[name=imagefile').next().html(language.gettext('label_choosefile'));
+                $('[name=imagefile').parent().next('input[type="button"]').fadeOut('fast');
+                if (prefs.usertheme) {
+                    loadBackgroundImages(prefs.usertheme);
+                } else {
+                    loadBackgroundImages(prefs.theme);
+                }
             });
         },